Undo previous changes, move i18nProvider inside App.js
